Created a Directory component
